+++ Angular-Front-End/app-front-end/src/app/models/role.model.ts
@@ -1,0 +1,5 @@
+export enum Role {
+    User = 'USER',
+    Editor = 'EDITOR',
+    Admin = 'ADMIN'
+}