+++ Angular-Front-End/app-front-end/src/app/services/authentication/auth-guard.service.ts
@@ -12,6 +12,12 @@ export class AuthGuard implements CanActivate {
     canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
         const currentUser = this.authenticationService.currentUserValue;
         if (currentUser) {
+            if(route.data.roles && route.data.roles.indexOf(currentUser.role) === -1){
+                // role not authorized, so redirect to home page
+                this.router.navigate(['/']);
+                return false;
+            }
+
             // authorised so return true
             return true;
         }