+++ Services/ComunityCatalogService/src/main/java/io/App/CommunityService/communityComponent/Role.java
@@ -1,0 +1,5 @@
+package io.App.CommunityService.communityComponent;
+
+public enum Role {
+	ADMIN, EDITOR, USER
+}